+++ Cosette/Engine/Ai/Search/SearchStatistics.cs
@@ -61,6 +61,8 @@ namespace Cosette.Engine.Ai.Search
         public int NullMovePrunes { get; set; }
         public int StaticNullMovePrunes { get; set; }
         public int FutilityPrunes { get; set; }
+        public int QSEEPrunes { get; set; }
+        public int QFutilityPrunes { get; set; }
 
         public Move[] PrincipalVariation { get; set; }
         public int PrincipalVariationMovesCount { get; set; }